split main function and test function into their own files
small modifications to make executable and test compile properly
